refactor(frontend): extract quote detail helpers
