+++ frontend/package.json
@@ -12,6 +12,7 @@
     "test:quote-workflow": "node scripts/quote-workflow-roundtrip.test.mjs",
     "test:quote-detail-mapping": "node scripts/quote-detail-mapping.test.mjs",
     "test:quote-edit-hydration": "node scripts/quote-edit-hydration.test.mjs",
+    "test:quote-detail-helpers": "node scripts/quote-detail-helpers.test.mjs",
     "test:crm-quote-prefill": "node scripts/crm-quote-prefill.test.mjs",
     "test:domestic-service-scope": "node scripts/domestic-service-scope.test.mjs",
     "test:spot-finalization": "node scripts/spot-finalization.test.mjs",
